dim_database

Module media

Structs

InsertableMedia

Struct which represents a insertable media object. It is usually used only by the scanners to insert new media objects. It is the same as Media except it doesnt have the id field.

Media

Media struct that represents a media object, usually a movie, tv show or a episode of a tv show. This struct is returned by several methods and can be serialized to json.

UpdateMedia

Struct which is used when we need to update information about a media object. Same as InsertableMedia except library_id cannot be changed and everything field is a Option<T>.