dim_database::query_ext

Trait QueryExt

pub trait QueryExt<'a, DB: Database> {
    // Required method
    fn bind_all<Vs, V>(self, values: Vs) -> Self
       where V: Send + Encode<'a, DB> + Type<DB> + 'a,
             Vs: IntoIterator<Item = V>;
}

Trait contains some extensions for sqlx.

Required Methods

fn bind_all<Vs, V>(self, values: Vs) -> Self
where V: Send + Encode<'a, DB> + Type<DB> + 'a, Vs: IntoIterator<Item = V>,

Method which allows you to bind several values in one go. This method will accept any container which can be turned into an Iterator of values.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.